refactor: extract prompt budget policies from promptBuilder into promptBudget
Move PromptSectionName, PromptBudgetPolicy, PromptBudgetPolicyKey types,
CODEX/CLAUDE_PROMPT_BUDGET_POLICIES constants, and buildPromptBudgetPolicy,
estimateTokenCount, estimateTokenRange functions into dedicated promptBudget.ts.
Reduces promptBuilder.ts from 1,424 to 1,100 lines. Backward-compatible via
export * re-export.
